refactor: move CMSImage and CMSVideo into SharedComponents
