+++ src/components/BioModal/BioModal.tsx
@@ -11,7 +11,7 @@ import {
 import AspectRatio from "@/components/SharedComponents/AspectRatio";
 import Box from "@/components/SharedComponents/Box";
 import IconButton from "@/components/SharedComponents/Button/IconButton";
-import CMSImage from "@/components/CMSImage";
+import CMSImage from "@/components/SharedComponents/CMSImage";
 import Flex from "@/components/SharedComponents/Flex";
 import Grid, { GridArea } from "@/components/Grid";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
