+++ src/components/NewButtons/HomepageTabImageButton/HomePageTabImageButton.tsx
@@ -17,7 +17,7 @@ import typography, { FontVariant } from "@/styles/utils/typography";
 import UnstyledButton, {
   UnstyledButtonProps,
 } from "@/components/UnstyledButton";
-import CMSImage from "@/components/CMSImage/CMSImage";
+import CMSImage from "@/components/SharedComponents/CMSImage/CMSImage";
 import Flex from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
 import TagPromotional from "@/components/TagPromotional";
