+++ src/components/Posts/PostPortableText/PostImageWithAltText.tsx
@@ -1,7 +1,8 @@
 import { PortableTextComponentProps } from "@portabletext/react";
 
 import { Image } from "../../../common-lib/cms-types";
-import CMSImage from "../../CMSImage";
+
+import CMSImage from "@/components/SharedComponents/CMSImage";
 
 const PostImageWithAltText = (
   props: PortableTextComponentProps<{ asset: Image["asset"] }>,
