+++ src/components/Posts/PostPortableText/PostVideo.tsx
@@ -1,8 +1,8 @@
 import { PortableTextComponentProps } from "@portabletext/react";
 
 import { Video } from "../../../common-lib/cms-types";
-import CMSVideo from "../../CMSVideo";
 
+import CMSVideo from "@/components/SharedComponents/CMSVideo";
 import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
