+++ src/components/SharedComponents/CMSImage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./CMSImage";
 export { sanityClientLike } from "./sanityImageBuilder";
+export type { CMSImageProps } from "./CMSImage";
