+++ src/components/SubjectIconBrushBorders/SubjectIconBrushBorders.tsx
@@ -1,12 +1,13 @@
 import { FC } from "react";
 
-import CMSImage from "../CMSImage";
-import { CMSImageProps } from "../CMSImage/CMSImage";
 import { getSubjectIconAsset } from "../../image-data";
 import Svg from "../Svg";
 import { OakColorName } from "../../styles/theme/types";
 import TagPromotional from "../TagPromotional";
 
+import CMSImage, {
+  CMSImageProps,
+} from "@/components/SharedComponents/CMSImage";
 import Flex from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
