+++ src/components/pages/CurriculumInfo/tabs/OverviewTab/OverviewTab.tsx
@@ -11,8 +11,8 @@ import Typography from "@/components/Typography/Typography";
 import { CurriculumOverviewMVData } from "@/node-lib/curriculum-api-2023";
 import { CurriculumOverviewSanityData } from "@/common-lib/cms-types";
 import { CurriculumSelectionSlugs } from "@/pages/teachers/curriculum/[subjectPhaseSlug]/[tab]";
-import CMSImage from "@/components/CMSImage";
-import CMSVideo from "@/components/CMSVideo";
+import CMSImage from "@/components/SharedComponents/CMSImage";
+import CMSVideo from "@/components/SharedComponents/CMSVideo";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 
 export type OverviewTabProps = {
