+++ src/components/pages/LandingPages/LandingPageTextAndMedia.tsx
@@ -7,8 +7,8 @@ import { TextAndMedia } from "@/common-lib/cms-types";
 import { getLinkHref } from "@/utils/portableText/resolveInternalHref";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Card from "@/components/SharedComponents/Card";
-import CMSImage from "@/components/CMSImage";
-import CMSVideo from "@/components/CMSVideo";
+import CMSImage from "@/components/SharedComponents/CMSImage";
+import CMSVideo from "@/components/SharedComponents/CMSVideo";
 import Flex from "@/components/SharedComponents/Flex";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders";
 import { LI } from "@/components/Typography";
