+++ src/pages/about-us/partners.tsx
@@ -9,13 +9,14 @@ import AboutContactCard from "../../components/AboutContactCard";
 import AboutIntroCard from "../../components/AboutIntoCard/AboutIntroCard";
 import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
 import AboutUsSummaryCard from "../../components/pages/AboutUs/AboutUsSummaryCard";
-import CMSImage from "../../components/CMSImage";
-import { CMSImageProps } from "../../components/CMSImage/CMSImage";
 import { SpacingProps } from "../../styles/utils/spacing";
 import Illustration from "../../components/Illustration";
-import { getSizes } from "../../components/CMSImage/getSizes";
 import getPageProps from "../../node-lib/getPageProps";
 
+import { getSizes } from "@/components/SharedComponents/CMSImage/getSizes";
+import CMSImage, {
+  CMSImageProps,
+} from "@/components/SharedComponents/CMSImage";
 import AspectRatio from "@/components/SharedComponents/AspectRatio";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Flex from "@/components/SharedComponents/Flex";
